refactor(ide): scripts
Decided not to remove duplication regarding the patch to apply to
configuration file in scripts. The complexity is not worth the benefit.
Anyway, if one side change (authserver or worldserver build) and not the
other (ide scripts) things probably may go weird and be noticed quite
quickly by the developer (hopefully).